Improve ntp klib accuracy and correction rate
Instead of using a single rate to correct both frequency and offset errors,
these changes use a least squares linear regression to estimate the values
separately using a relatively large number of data points, where the slope
represents the frequency error and the intercept equals the offset. With this
method of calculating frequency error, it the ntp klib can now track
the reference clock with sub-millisecond per second accuracy. The frequency
error gets applied to all clocks except for MONOTONIC_RAW (by definition).
The slewing adjustments are applied only to the realtime clock. With the
separation of error components, the offset error can be much more
aggressively slewed away, correcting offsets of over a second within a single
query interval. Older samples are adjusted for offset changes so that the
linear regression is not influenced by the error introduced by the slewing
frequency. The slew rate is inversely proportional to the size of the offset
so that errors in slewing are reduced as the offset size approaches zero.
All computations are performed with a 32-bit fractional signed
fixed point type which has enough precision to keep errors in the
nanoseconds range. To prevent a non-sensical linear regression
from setting the clock to a bad value, the estimated frequency
is clamped to a reasonable limit, and enough bad regression
computations will result in resetting regression state to discard
any bad values contributing to the problem.

Additional sanity checks have been added for the ntp response packet
based on recommendations by the NTP RFC. These mostly involve
discarding packets as the result of server or network misbehavior.

Time related functions which read vdso state now check a generation
number to ensure that all vdso_dat reads are consistent with each
other.

Two new manifest options have been added: ntp_max_slew_ppm which
determines the offset slewing rate and ntp_max_freq_ppm which controls
the maximum frequency error rate.

The initial polling interval is set to the minimum at startup in order to calculate
initial frequency and offset errors so they will be corrected within the
first few minutes of runtime.

In order to help with debugging, a new print format %f has been added for
printing s64 fixed point values. This is similar to %T for timestamps
but with more decimal values printed.