virtio-rng support (#1732)
This introduces support for the virtio entropy device. When available, this
device will serve as the main source of entropy for the random number
generator. A hardware entropy source (e.g. rdseed and rdrand on x86-64) serves
as a fallback option when a virtio entropy device is not found.

Entropy data is stored in a double-buffered fashion with the intent to keep
entropy available for randomize functions without the need to block. Should an
underrun in entropy data occur, the aforementioned fallback option is used to
fulfill the request.

Other changes include adding a mutex to guard chacha20 usage and splitting
large getrandom(2) requests into smaller chunks to avoid starving other work.