Azure: Handle multi-packet goalstate response from wireserver
When the Azure cloud init issues a GET request for the goal state in order to
report status, the XML reponse from the wireserver may be large enough to
require multiple packets. This change fixes a bug where the response
was assumed to fit in a single packet, so the request parser was never
called unless the response was small enough. The http parser is moved
to the receive handler from the packet handler so that parser state is
preserved between packets. The response parser now closes the connection
when any complete http response is received.