Syscall stats: properly initialize thread syscall time
This change fixes an issue that was causing the execution time of
the first syscall executed by a newly created thread to be computed
incorrectly, due to the syscall_time field of the thread struct not
being initialized at syscall entry. The fix moves the
initialization of this field from count_syscall (which is executed
at syscall exit) to syscall_handler (which is executed at syscall
entry).