count_syscall_save(): properly handle multiple consecutive calls
The count_syscall_save() function can be called multiple times
without count_syscall_resume() being called in-between; for
example, thread_sleep_interruptible() calls count_syscall_save()
and then enters the runloop; then, if a contextual closure is
dequeued from a task queue, the current syscall context is paused,
which triggers another call to count_syscall_save().
If count_syscall_save() is called two consecutive times, at the
second call the syscall_enter_ts field of the thread struct is 0;
this change adds a check for a zero value, in order to prevent an
absolute timestamp value to be added to the syscall_time thread
struct member.