x86 restore_ucontext(): don't trust segment selector values in uctx
A signal handler can modify the ucontext of a thread, including
segment selectors in the cs and ss fields. If invalid selector
values are copied from the ucontext to a thread frame when the
signal handler returns, a general protection fault will occur
during execution of the iretq instruction to resume the thread; in
this state, the GS base register contains a userspace value
(because the swapgs instruction has already been executed in
`frame_return`) but the CPU is still in kernel mode, therefore the
interrupt handler for the general protection fault will use the GS
base register as if it contained a kernel value (i.e. a cpuinfo
pointer); this poses a security risk because a compromised
application could craft a rogue cpuinfo struct and place a pointer
to it in the GS base register, which will be used by the kernel.

This changes fixes the above issue by modifying restore_ucontext()
so that CS and SS values in a thread frame are always valid,
regardless of any changes made to them by the user process.