setitimer/alarm: always use CLOCK_ID_MONOTONIC
All itimer types (including ITIMER_REAL) should use the monotonic
clock and should not be affected by any discontinuous changes to
the real-time clock. This is true for alarm(2) as well, since
internally it uses an ITIMER_REAL itimer.
The only difference between ITIMER_REAL and ITIMER_PROF is that the
latter should account only for actual CPU time consumed by the
process, but currently the kernel treats it as ITIMER_REAL; code
comments have been revised to clarify this point.