move scheduling of thunks to async_apply() and async_apply_bh()
This change creates the async_apply() and async_apply_bh() functions for
scheduling thunks to the runqueue and bhqueues, respectively. Previously,
thunks were directly enqueued to the scheduling queue, which proved to be
error-prone as return values were often unchecked, leading to queueing
failures being undetected. The selection between enqueue() vs
enqueue_irqsafe() variants is also risky, as a (non-irqsafe) enqueue()
competing with an enqueue from an interrupt handler could potentially lead to
a hang. The semantics for the two queues have been clarified such that
scheduling to the runqueue is not allowed from interrupt handling; scheduling
from interrupt level must occur only to the bhqueue. This interface also hides
the implementation details of the scheduling queues, thus making it easier to
revise these details without affecting much of the calling code.

All async_apply*() functions are now void returns, indicating that such
scheduling should never fail. Internally, the functions simply assert that
enqueues succeed, but in the future they could possibly lead to alternate
paths to mitigate the limit (such as suspending the execution context, if
possible, while queues are drained and resources are made available). It may
also be desirable to adopt a lock-free queue that is extendable, avoiding the
need for fixed queue sizes altogether.

Within the scheduler, the queue sizes set on initialization are now defined in
config.h, and the async_queue_1 size has been expanded to make room for large
bursts of enqueues, such as those that occur while servicing virtqueues.