clock_reset_rtc(): fix timer reordering and concurrency issues
When the real-time clock undergoes a discontinuous change,
modification of timer expiry values and reordering of active timers
must be done with the timerqueue locked (this will allow safely
chagning the real-time clock from multiple sources, such as
syscalls and NTP); in addition, the RTC offset value in the VDSO
structure must be reset (with the timerqueue locked) before
reordering the timers, because timer expiry calculations done
during the reordering should use the updated offset.
This commit adds the timer_adjust_begin() and timer_adjust_end()
functions, which allow executing multiple operations (such as
resetting the RTC offset and adjusting the timer expiry values)
with the timerqueue locked, and reorder the timers before unlocking
the timerqueue.