Datagram sockets: implement MSG_TRUNC option
When an input datagram is truncated during execution of recvmsg or
recvmmsg, the MSG_TRUNC flag is set in the msg_flags member of
struct msghdr. If this flag is specified in the recvmsg or recvmmsg
arguments, the length value returned for a truncated datagram is
the length of the original datagram instead of the truncated
length.