Improve robustness of fs_full test
This test was failing frequently with a false test failures. The
write_blocks function is reworked with a smarter retry so that enough
time elapses to allow the log flush to finish freeing old logs before
giving up on writes. The write test when the filesystem is full now
allows writing up to bfree blocks instead of assuming that any write
will fail. The file remove and free space check now syncs and waits
after the remove is called to allow time for the free space to be
updated before calling stat.