vm_exit(): move common code to kernel/init.c
With this change, the vm_exit() function contains
non-platform-specific code to be executed just before terminating
the kernel, while 2 new platform-specific functions (vm_shutdown
and vm_reset) implement the actual shutdown and reset of a VM.
Handing of the reboot_on_exit option is being modified so that this
option only takes effect if the exit code is not the expected one
(i.e. if either the application terminates with error, or the
kernel crashes), because this was the original intent behind this
feature, and rebooting unconditionally makes it impossible to
cleanly stop a running instance (e.g. with `ops instance stop`).