vm_exit(): add idle_on_exit option
This change introduces a new option, called "idle_on_exit", which
when specified in the manifest keeps a VM running (with all VCPUs
halted) after the user program exits successfully.
The implementation of the machine_halt() function for aarch64 and
risc-v is being fixed so that it actually halts the CPU and does
not return.
Generation of the broadcast shutdown IPI (which is only needed when
the idle_on_exit option takes effect) is being removed from where
it is not needed.