Deallocate extremely large sg lists
When an sg_list is deallocated it goes onto a free list instead of released
to the heap. This change reallocates the sg_list buffer if it is larger than
a threshold specified in config.h so that large amounts of memory are not
tied up in freed sg_lists.