Pagecache: commit dirty pages in chunks
Previously, pagecache_commit_dirty_node() would issue a single storage
request for all dirty pages of a file. If there were enough dirty pages and
free memory was low, this would panic the system trying to allocate
resources for the request. This change makes pagecache_commit_dirty_node()
submit storage i/o in smaller chunks where the chunks are serialized so that
memory usage is minimized when there are a large number of dirty pages.

The completion for the dirty commit was attached with
pagecache_finish_pending_writes(), but this will run the completion too soon
when the dirty pages are flushed in chunks. This is modified so the completion
is now passed to pagecache_commit_dirty_node().

If pagecache_commit_dirty_node() is called again while a previous commit
is still running, the commit is queued on the node and will be run after the
previous commit finishes.