rangemap_reinsert: roll back to old range and reinsert on failure
If (for e.g.) the new heap range requested by brk was previously
mapped by mmap then rangemap_insert will fail. When this happens we
want to leave the rangemap how we found it. Otherwise it effectively
unmaps the heap.

So this reinserts the original range if the new one fails. Other
callsites of rangemap_reinsert all assert that it passes. So this
change should not effect them.