Pagecache: fix handling of concurrent page writes and reads
The existing code in pagecache_write_sg() is trying to prevent a
read of a newly allocated page in the middle of a write (to the
same page) from triggering a fetch from disk by setting the page
state to WRITING. However, doing so causes a page read occurring in
the middle of a write to return garbage data.
This change fixes the issue by checking for any ongoing page reads
in the write bottom half, and postponing the completion of the
write so that the data being written will overwrite the data
fetched by the ongoing read requests.
The change_page_state_locked() function has been changed so it can
handle a state transition from ALLOC to DIRTY.