Pagecache: fix page refcounting and races with page eviction
Having a page refcount that can be reserved and released without
the page state lock is racy with respect to the pagecache drain
functionality, as illustrated in the following example:
- cpu0: a page is being evicted, its refcount is decremented from
1 to 0
- cpu1: a page read (or write) is requested, which increments the
  refcount from 0 to 1, then acquires the state lock, verifies that
the page is not in the FREE state and releases the lock
- cpu0 acquires the state lock, changes the page state to FREE and
  releases the state lock
- cpu1 continues handling the read (or write) request on a page in
  the FREE state whose backing memory has been deallocated
To fix the above issue, this commit changes the page refcount
handling so that the refcount is only incremented and decremented
with the page state lock held. For read operations, a conventional
refcount is still needed, because the consumer of a read SG list
may access the read data at an arbitrary time after the data has
been read, and the relevant pages must not be freed during this
time; the read refcount is tied to the generic refcount so that the
generic refcount is incremented and decremented when the read
refcount changes from 0 to 1 and from 1 to 0, respectively.
The touch_or_fill_page_nodelocked() function has been changed so
that it increments the page refcount before releasing the state
lock.
The pagecahe drain code now can evict pages without needing a
temporary evict list.
When a pagecache node is deallocated, the refcount of its pages is
released only for non-evicted pages: this fixes a
'refcount_release: invalid count 0' error occurring when a node
with evicted pages is deallocated.
The write code was missing a page refcount release after write
completion; this is now fixed by adjusting the refcount depending
on whether a page is dirty. The commit completion closure
decrements the refcount unless the page state is set back to DIRTY
due to a write error.
In pagecache_node_fetch_internal(), the page handler closure is
invoked after verifying that a page is not in the FREE state.