NTP: add ability to configure multiple servers
With this change, it is possible to configure the NTP klib with
multiple NTP servers, by inserting an `ntp_servers` array in the
root tuple of the manifest, where each array element follows the
syntax '<addr>[:<port]'; <addr> can be an IP address or a fully
qualified domain name; if the optional <port> is not present, the
default NTP port 123 is used.
Example configuration with 2 servers:
`ntp_servers:[us.pool.ntp.org 12.34.56.78:1234]`
When multiple servers are configured, the NTP klib starts by using
the first server, and switches to the next server (in a round robin
fashion) if no valid response is received from the current server.