Kibs: ntp: add IPv6 support
This change allows the NTP klib to work on VMs with IPv6 internet
connectivity.
In the `<addr>[:<port>]` syntax that specifies an NTP server,
`<addr>` can be a numeric IPv6 address; in this case, it must be
enclosed in square brackets, as per RFC 3986, and the
`<addr>[:<port>]` string must be enclosed in double quotes in order
to be parseable by the tuple parser.
Example configuration with 2 IPv6 servers:
`ntp_servers:["[2610:20:6f97:97::4]" "[2610:20:6f97:97::5]:1234"]`