Objcache: add locking option and drain functionality
This change introduces a new heap sub-type, called caching_heap,
that exports an additional callback to drain memory that has been
cached. The objcache heap is now a caching_heap.
In addition, the objcache heap now implements locking, which allows
safe use of the drain functionality in concurrently accessed heaps
without the need for an external lock.
Memory cleaner instances have been added in various places where an
objcache is used, so that caching heaps can be drained when the
system is low on memory.

Partially addresses #1494.