Klibs: ntp: probe all configured servers at startup
With this change, the ntp klib starts in a "probe mode" in which it
tries to communicate with all configured NTP servers; the first
server that replies with valid data is selected as the current
server. This prevents non-reachable or invalid servers from
delaying synchronization of the VM system time if there is at least
one valid and reachable server.
Once a server is selected, any data received from other servers is
discarded. If data received from the current server later becomes
invalid, the ntp klib returns to probe mode.
Since communication may happen in parallel with multiple servers,
the last_originate_time and last_transmit_time fields, used for
sanity checks, are being moved from the global ntp struct to the
struct associated to a given server.