Scheduler: implement fair scheduling between user threads
This change introduces a new scheduling algorithm that assigns CPU
resources to a task based on how much CPU time the task used in
the past. It uses a priority queue where tasks are ordered based on
their CPU time: when a task is enqueued, it is positioned according
to the time it has been running since it was last scheduled; when a
task is dequeued, its CPU time is subtracted from the CPU time of
any other tasks in the queue, so that their priority in the queue
is increased; this is implemented by storing a reference CPU time
in the queue (min_runtime struct member), which prevents
long-running tasks from being starved by newly created tasks.
This algorithm is applied for scheduling user threads, so
that CPU time is distributed fairly between all runnable threads.