Unix fault handler: orphan syscall context for abandoned syscalls
When an unhandled page fault occurs in a syscall context, the
syscall is abandoned and a SIGSEGV is raised. The syscall context
should not be reused by other threads until the offending thread
runs the signal handler, otherwise if another thread executes a
syscall, the assert(sc->context.refcount.c == 1) in the syscall
handler fails.
This commit fxes the above issue by adding a call to
orphan_syscall_context() in the Unix fault handler, so that the
syscall context is not reused by any threads that might run on the
same CPU before the signal handler of the offending thread. Since
orphaning a context releases its refcount, a call to
context_release_refcount() is being removed from the arch-specific
exception handlers.