sendmsg/recvmsg socket callbacks: allow calls from bottom half
With this change, the sendmsg and recvmsg callbacks take two
additional parameters that allow them to be called not only from
the top half of a syscall, but also from a bottom half context.
This expands the functionality of the existing sendmmsg syscall to
cover not only network sockets, but any type of sockets that
implements the sendmsg syscall.
For network sockets, the code that handles data transmission and
reception has been refactored to use iovec structures instead of a
single continuous buffer; this allows avoiding the additional
buffer allocation and copying that was done for sendmsg, sendmmsg,
and recvmsg.