pagecache: fix committing of nodes with empty `dirty` rangemap
pagecache_commit_dirty_node() copies any dirty ranges in a node to
a buffer and removes them from the `dirty` rangemap; then, it
triggers the actual committing of the dirty ranges. If
pagecache_commit_dirty_node() is called when the rangemap is empty,
the completion closure is invoked immediately even if the
committing is ongoing, which is incorrect and can cause issues such
as premature deallocation of fsfiles that have not been
synchronized to disk.
In addition, the buffer is not properly deallocated when the
rangemap is empty.
This commit fixes the above issues.