lwIP options: increase TCP send buffer size and TCP window size
Since lwIP implements TCP window scaling, the window size is not
limited to a maximum of 65535. Increasing the TCP send buffer size
to 0x34000 (same value as default setting on Linux) has been shown
to improve Warp (https://github.com/minio/warp) benchmark results
when performing HTTP GET operations against a minIO server.

There is a known issue in the loopback network interface code that
causes TCP data re-transmitted after a missing acknowledge to not
wake up the receiving end of a TCP connection; this is due to the
fact that a TCP re-trasmisssion (which is done by the TCP timer
function) does not involve any code in src/net/netsyscall.c, which
then cannot trigger a poll of the loopback interface. This issue
manifests itself when incoming data is discarded due to the buffer
of the receiving end of a TCP connection being full, which can
happen if this buffer is smaller than the send buffer (and also if
the number of buffered packets exceeds SOCK_QUEUE_LEN). For this
reason, the receive buffer size in the netsock runtime test is now
kept at its default value (which matches the send buffer size), and
the value of SOCK_QUEUE_LEN is increased; these changes allow the
netsock runtime test to pass.