virtIO: implement device interrupt suppression
This commit adds support for the VIRTIO_F_RING_EVENT_IDX feature,
which allows a virtIO driver to indicate to the device when it
wants to be notified of an update to used buffers, and similarly
allows a virtIO device to indicate to the driver when it wants to
be notified of an update to available buffers. By avoiding unneeded
device interrupts and driver notifications, the number of VM exits
and hypervisor syscalls (e.g. to execute KVM_SIGNAL_MSI ioctls) is
minimized, which decreases CPU load on the host.
Enabling this virtIO feature in the virtio_net driver has been
shown to improve the speed of large TCP transfers from guest to
host.