virtIO: implement polling mode
When a virtIO queue is in polling mode, the driver does not want to
be notified (via an interrupt) when a virtIO buffer in the queue
has been used by the device. Avoiding unnecessary interrupts
decreases the CPU load on the host machine.
Polling mode is now enabled in the TX queue of the virtio_net
driver, because the message completion routine for a used TX buffer
just deallocates the buffer, which is not a time-critical operation
and as such can be deferred. This has been shown to improve the
speed of large TCP transfers from guest to host.