fsync/fdatasync: avoid TFS log flush when not needed
When fsync or fdatasync is called on a file descriptor for a
regular file, a flush of the TFS log to the storage device is only
needed if the file metadata have changed (for fdatasync, this
applies only to metadata needed in order to allow a subsequent data
retrieval to be correctly handled). This commit adds some status
flags to the fsfile struct to track whether file metadata have
changed since the last flush, and to avoid flushing the log when
metadata have not changed (note: currently, log flushes triggered
outside the context of fsync/fdatasync syscalls, such as periodic
flushes, do not affect the fsfile status flags).
The filesystem_sync() function has been removed in favor of using
filesystem_flush(), because there was duplicated code between these
functions, and each filesystem_sync() call was triggering two calls
to pagecache_sync_volume(). Similarly, filesystem_sync_node()
(which was triggering an unnecessary call to pagecache_sync_volume)
is being replaced by fsfile_flush(). In order to ensure that the
completion of an fync/fdatasync syscall runs in the appropriate
syscall context, calls to apply() have been replaced by calls to
async_apply_status_handler() in the TFS code and the storage
drivers that implement the flush operation.
filesystem_flush() and fsfile_flush() are implemented so that the
page cache is flushed before the TFS log: this is necessary to
prevent incorrect data from being associated to a file if a VM is
stopped after a log flush but before a page cache flush.

fsync() on a directory file descriptor now properly flushes the TFS
log, because this syscall can be used to ensure newly created
directory entries are saved to the storage device. fsync() and
fdatasync() on a symlink file descriptor now return -EBADF, for
consistency with the Linux kernel.