open(2): add support for O_SYNC and O_DSYNC flags
When a regular file is opened with O_SYNC, by the time write(2)
(or similar) returns, the output data and associated file metadata
have been transferred to the underlying hardware. The O_DSYNC flag
is similar, but it only flushes file metadata if they are needed to
correctly handle a subsequent data retrieval.
This commit implements the above functionality by calling
fsfile_flush() as appropriate in the completion functions of file
write operations.