Add devicetree blob parsing support
This change adds basic support for parsing and reading from the devicetree
blob (DTB) found in RISC-V and ARM machines. The DTB contains information
about a machine and its devices and how to access or configure them. Parsing
of the DTB does not require any special kernel features, so the devicetree
code is located in its own directory so it can be used by the kernel or by
user tools.

The ranges of memory in a machine are specified only in the DTB. Nanos has
previously been assuming a fixed size memory for RISC-V and ARM VMs, so this
change now allows Nanos to discover the actual amount of memory on a machine.

The devicetree_dump() function will dump the contents of the DTB, given its
address.

In order to accommodate the limited boot environment in the kernel when the
DTB will normally be accessed, the code uses only stack memory, no heap.
Any extensive processing of the DTB should cache the tree structure to
make navigation more efficient. However, that is not included in this change.