mutex: do not save current cpuinfo across context_suspend
mutex_lock_internal() was saving a local copy of the current cpuinfo prior to
calling context_suspend(). This copy is later passed to calls mcs_unlock()
after the context resumes. Since a context may be suspended on one CPU and
resumed on another, this could potentially lead to use of a cpuinfo that does
not belong to the current CPU. This change replaces such uses of the
local copy with calls to current_cpu().