intern(): do not use random_u64()
The call to random_u64() from within intern() leads to the chacha20 mutex
being locked. This is problematic if such a call were to occur from within an
interrupt handler (e.g. on a call to tprintf()) or in any other situation
where a context suspend would not be appropriate. To avoid this, a simple
hashing function has been borrowed from the x86 stage2 code to use in place of
the call to random_u64().