implement multiple IPI vectors for RISC-V
This change adds support for multiple pending IPI vectors in riscv64 interrupt
support. A bitmask word of pending vectors is kept in each cpuinfo_machine. On
a call to send_ipi(), the corresponding bit is atomically set in the
word. When trap_interrupt() handles a software interrupt (type TRAP_I_SSOFT),
the word is atomically swapped with zero, and handlers for each set
bit (pending vector) are invoked. The "flush_handler" kludge is removed, and
calls to send_ipi() for the shutdown vector are restored in trap_interrupt()
and trap_exception(). This brings the riscv64 IPI implementation in line with
those of other architectures.