Network drivers: add support for gVNIC adapter
The gVNIC (also known as GVE, Google Virtual Ethernet) network
adapter is used on Google Cloud instances, and is the only adapter
type supported on t2a (aarch64) instances.
This change implements a gVNIC driver and enables it in the aarch64
kernel, which allows Nanos images to run with networking support on
Google Cloud ARM instances.