tracelog file writer
This adds an option to the tracelog facility for recording trace entries to a
file on the root partition. If a 'file' attribute is found in the tracelog
manifest tuple, a file with the name of the corresponding value is created in
the root directory and file tracing is enabled. For ease of debugging, this
attribute may be added to the manifest at build time by using the
TRACELOG_FILE build option, e.g.:

$ make TARGET=hw TRACE=tracelog DEBUG=sched TRACELOG_FILE=foo run

The trace data from execution will be left in the file named 'foo' in the root
disk image. The tracelog file may be copied out of the image using the
'copylog' make target, which will leave a copy of the file in the output
directory.

Note that enabling the file writer will start a periodic timer to collate and
flush trace data to disk. This is to ensure that pending trace data is written
in a timely fashion in case the system is unable to cleanly shut down. A
shutdown completion is registered to flush pending data prior to termination.

The documentation in 'doc/tracelog.txt' has been updated accordingly, and mkfs
has been updated to allow multiple instances of the '-t' commandline option.