Unix: create /proc/self/exe even when /proc/self directory exists
A given image may be created with a /proc/self directory. The
existing code is skipping creation of the /proc/self/exe symlink
when the /proc/self directory exists; this commit changes the code
so that the symlink is always created when not present, regardless
of the presence of the containing folder.