Klibs: tls: fix handling of connection shutdown
If the application layer of a TLS connection requests that the
connection is shut down from within the input buffer handler, by
the time the handler (as called by tls_in_handler) returns, the
connection may have been shut down, and the tls_conn struct may
have been deallocated. To prevent tls_in_handler from being called
again due to further network input packets that may have already be
queued for this connection, tls_in_handler must return true. In
order for tls_in_handler to safely access the tls_conn struct to
determine whether the connection is still active, a reference count
is being added to this struct.
This change fixes a crash that may occur when the cloud_init klib
tries to download a file from an HTTPS server that uses chunked
transfer encoding (which is not supported).