Klibs: connect to remote server after asynchronous DNS resolution
When the IP address of the server with which a klib must
communicate is unavailable (because either it has never been
resolved from a hostname, or its entry in the DNS cache has
expired), the dns_gethostnyname() function completes asynchronously
via a user-supplied callback. Some klibs are communicating with the
remote server only when the above function completes synchronously,
i.e. when there is a valid entry in the DNS cache. This creates
problems when the time interval between successive calls to
dns_gethostnyname() is longer than the TTL of the server's DNS
entry (as can happen with the radar and cloudwatch klibs), and may
prevent the syslog klib from sending to the syslog server messages
generated when there is no valid DNS entry for the server.
This commit fixes the above issue by adding code in the DNS
callback functions to connect to the remote server after an
asynchronous DNS resolution; since the callbacks are invoked with
the lwIP lock held, the radar and cloudwatch klibs use a thunk that
is applied asynchronously in order to be able to call functions
such as direct_connect() or tls_connect().