Klibs: strip debug information from klib binaries
This makes klib binary files smaller, thus allowing more klibs to
fit in the relatively small boot partition. Binaries with debug
information are still generated in the build process, and are kept
as separate files, with .dbg extension.
The CircleCI script is being amended so that generated tarballs do
not include files with debug information.