Unix special files: add support for major and minor number
A special file can now be associated to a major and minor number
(which are encoded in the filesystem metadata). The stat, fstat and
lstat syscalls retrieve the major and minor number from the
filesystem metadata and put them in the `st_rdev` field of struct
stat.
This feature is useful to run programs that check the `st_rdev`
field against known values, such as programs using the Nvidia GPU
libraries.