Context pause/resume: invoke callbacks even during kernel shutdown
The pause and resume context callbacks in the kernel context type
adjust the context reference count. If the reference count of a
kernel context that is about to be run is not bigger than 1, the
kernel_context_return() function triggers an assertion failure.
Not invoking the pause and resume callbacks during kernel shutdown
can cause a panic when a kernel context is dequeued from a bottom
half queue; e.g. when an image with the Nvidia GPU klib is run on
an instance that doesn't a have a GPU, an assertion failure is
triggered when a kernel thread from the Nvidia klib is about to be
run during kernel shutdown.
This change fixes the above issue by invoking the pause and resume
context callbacks regardless of the shutdown status.